+++ packages/smarthr-ui/src/components/DatePicker/DatePicker.tsx
@@ -23,7 +23,7 @@ import { tv } from 'tailwind-variants'
 import { useOuterClick } from '../../hooks/useOuterClick'
 import { textColor } from '../../themes'
 import { Calendar } from '../Calendar'
-import { FaCalendarAltIcon } from '../Icon'
+import { FaCalendarDaysIcon } from '../Icon'
 import { Input } from '../Input'
 
 import { Portal } from './Portal'
@@ -31,7 +31,7 @@ import { parseJpnDateString } from './datePickerHelper'
 import { useGlobalKeyDown } from './useGlobalKeyDown'
 
 type ChangeLikeEvent = ChangeEvent | React.KeyboardEvent | MouseEvent
-type Props = {
+type AbstractProps = {
   /** input 要素の `value` 属性の値 */
   value?: string | null
   /** input 要素の `name` 属性の値 */
@@ -64,15 +64,18 @@ type Props = {
     other: { date: Date | null; formatValue: string; errors: string[] },
   ) => void
 }
-type OmitInputAttributes =
-  | keyof Props
-  | 'type'
-  | 'onKeyPress'
-  | 'onFocus'
-  | 'aria-expanded'
-  | 'aria-controls'
-  | 'aria-haspopup'
-type InputAttributes = Omit<ComponentPropsWithRef<'input'>, OmitInputAttributes>
+type Props = AbstractProps &
+  Omit<
+    ComponentPropsWithRef<'input'>,
+    | keyof AbstractProps
+    | 'type'
+    | 'onChange'
+    | 'onKeyPress'
+    | 'onFocus'
+    | 'aria-expanded'
+    | 'aria-controls'
+    | 'aria-haspopup'
+  >
 
 export const DEFAULT_FROM = new Date(1900, 0, 1)
 
@@ -93,7 +96,7 @@ const RETURN_NULL = () => null
 const ESCAPE_KEY_REGEX = /^Esc(ape)?$/
 
 /** @deprecated DatePicker は非推奨です。Input[type=date] を使ってください。 */
-export const DatePicker = forwardRef<HTMLInputElement, Props & InputAttributes>(
+export const DatePicker = forwardRef<HTMLInputElement, Props>(
   (
     {
       value,
@@ -441,12 +444,12 @@ export const DatePicker = forwardRef<HTMLInputElement, Props & InputAttributes>(
 const InputSuffixIcon = memo<{
   classNames: { inputSuffixLayout: string; inputSuffixWrapper: string; inputSuffixText: string }
   alternativeFormat: null | ReactNode
-  caretIconColor: ComponentProps<typeof FaCalendarAltIcon>['color']
+  caretIconColor: ComponentProps<typeof FaCalendarDaysIcon>['color']
 }>(({ classNames, alternativeFormat, caretIconColor }) => (
   <span className={classNames.inputSuffixLayout}>
     <span className={classNames.inputSuffixWrapper}>
       {alternativeFormat && <span className={classNames.inputSuffixText}>{alternativeFormat}</span>}
-      <FaCalendarAltIcon color={caretIconColor} />
+      <FaCalendarDaysIcon color={caretIconColor} />
     </span>
   </span>
 ))