+++ packages/smarthr-ui/src/components/Dialog/FormDialog/FormDialog.tsx
@@ -9,10 +9,10 @@ import { FormDialogContentInner, type FormDialogContentInnerProps } from './Form
 
 import type { DialogProps } from '../types'
 
-type Props = Omit<FormDialogContentInnerProps, 'titleId'> & DialogProps
-type ElementProps = Omit<ComponentProps<'div'>, keyof Props>
+type AbstractProps = Omit<FormDialogContentInnerProps, 'titleId'> & DialogProps
+type Props = AbstractProps & Omit<ComponentProps<'div'>, keyof AbstractProps>
 
-export const FormDialog: FC<Props & ElementProps> = ({
+export const FormDialog: FC<Props> = ({
   children,
   title,
   subtitle,
@@ -33,7 +33,7 @@ export const FormDialog: FC<Props & ElementProps> = ({
   decorators,
   id,
   isOpen,
-  ...props
+  ...rest
 }) => {
   const { createPortal } = useDialogPortal(portalParent, id)
   const titleId = useId()
@@ -55,7 +55,7 @@ export const FormDialog: FC<Props & ElementProps> = ({
 
   return createPortal(
     <DialogContentInner
-      {...props}
+      {...rest}
       isOpen={isOpen}
       ariaLabelledby={titleId}
       className={className}