+++ packages/smarthr-ui/src/components/Icon/FaIcon.tsx
@@ -130,6 +130,7 @@ import {
   FaList,
   FaListOl,
   FaListUl,
+  FaLocationDot,
   FaLock,
   FaLockOpen,
   FaMagnifyingGlass,
@@ -139,12 +140,14 @@ import {
   FaMicrosoft,
   FaMinus,
   FaMoneyCheckDollar,
+  FaMugHot,
   FaPaintbrush,
   FaPaperPlane,
   FaPaperclip,
   FaPen,
   FaPenToSquare,
   FaPencil,
+  FaPersonWalkingArrowLoopLeft,
   FaPiggyBank,
   FaPlus,
   FaPowerOff,
@@ -179,14 +182,12 @@ import {
   FaTags,
   FaThumbtack,
   FaToolbox,
-  FaTrash,
   FaTrashCan,
   FaTriangleExclamation,
   FaTruck,
   FaTruckRampBox,
   FaUpDown,
   FaUpRightAndDownLeftFromCenter,
-  FaUpRightFromSquare,
   FaUser,
   FaUserCheck,
   FaUserClock,
@@ -336,6 +337,7 @@ export const FaLinkSlashIcon = /*#__PURE__*/ generateIcon(FaLinkSlash)
 export const FaListIcon = /*#__PURE__*/ generateIcon(FaList)
 export const FaListOlIcon = /*#__PURE__*/ generateIcon(FaListOl)
 export const FaListUlIcon = /*#__PURE__*/ generateIcon(FaListUl)
+export const FaLocationDotIcon = /*#__PURE__*/ generateIcon(FaLocationDot)
 export const FaLockIcon = /*#__PURE__*/ generateIcon(FaLock)
 export const FaLockOpenIcon = /*#__PURE__*/ generateIcon(FaLockOpen)
 export const FaMagnifyingGlassIcon = /*#__PURE__*/ generateIcon(FaMagnifyingGlass)
@@ -345,12 +347,16 @@ export const FaMessageIcon = /*#__PURE__*/ generateIcon(FaMessage)
 export const FaMicrosoftIcon = /*#__PURE__*/ generateIcon(FaMicrosoft)
 export const FaMinusIcon = /*#__PURE__*/ generateIcon(FaMinus)
 export const FaMoneyCheckDollarIcon = /*#__PURE__*/ generateIcon(FaMoneyCheckDollar)
+export const FaMugHotIcon = /*#__PURE__*/ generateIcon(FaMugHot)
 export const FaPaintbrushIcon = /*#__PURE__*/ generateIcon(FaPaintbrush)
 export const FaPaperPlaneIcon = /*#__PURE__*/ generateIcon(FaPaperPlane)
 export const FaPaperclipIcon = /*#__PURE__*/ generateIcon(FaPaperclip)
 export const FaPenIcon = /*#__PURE__*/ generateIcon(FaPen)
 export const FaPenToSquareIcon = /*#__PURE__*/ generateIcon(FaPenToSquare)
 export const FaPencilIcon = /*#__PURE__*/ generateIcon(FaPencil)
+export const FaPersonWalkingArrowLoopLeftIcon = /*#__PURE__*/ generateIcon(
+  FaPersonWalkingArrowLoopLeft,
+)
 export const FaPiggyBankIcon = /*#__PURE__*/ generateIcon(FaPiggyBank)
 export const FaPlusIcon = /*#__PURE__*/ generateIcon(FaPlus)
 export const FaPowerOffIcon = /*#__PURE__*/ generateIcon(FaPowerOff)
@@ -386,7 +392,6 @@ export const FaTagsIcon = /*#__PURE__*/ generateIcon(FaTags)
 export const FaThumbtackIcon = /*#__PURE__*/ generateIcon(FaThumbtack)
 export const FaToolboxIcon = /*#__PURE__*/ generateIcon(FaToolbox)
 export const FaTrashCanIcon = /*#__PURE__*/ generateIcon(FaTrashCan)
-export const FaTrashIcon = /*#__PURE__*/ generateIcon(FaTrash)
 export const FaTriangleExclamationIcon = /*#__PURE__*/ generateIcon(FaTriangleExclamation)
 export const FaTruckIcon = /*#__PURE__*/ generateIcon(FaTruck)
 export const FaTruckRampBoxIcon = /*#__PURE__*/ generateIcon(FaTruckRampBox)
@@ -394,7 +399,6 @@ export const FaUpDownIcon = /*#__PURE__*/ generateIcon(FaUpDown)
 export const FaUpRightAndDownLeftFromCenterIcon = /*#__PURE__*/ generateIcon(
   FaUpRightAndDownLeftFromCenter,
 )
-export const FaUpRightFromSquareIcon = /*#__PURE__*/ generateIcon(FaUpRightFromSquare)
 export const FaUserCheckIcon = /*#__PURE__*/ generateIcon(FaUserCheck)
 export const FaUserClockIcon = /*#__PURE__*/ generateIcon(FaUserClock)
 export const FaUserIcon = /*#__PURE__*/ generateIcon(FaUser)
@@ -407,7 +411,3 @@ export const FaWindowMinimizeIcon = /*#__PURE__*/ generateIcon(FaWindowMinimize)
 export const FaWindowRestoreIcon = /*#__PURE__*/ generateIcon(FaWindowRestore)
 export const FaXmarkIcon = /*#__PURE__*/ generateIcon(FaXmark)
 export const FaYenSignIcon = /*#__PURE__*/ generateIcon(FaYenSign)
-
-export { WarningIcon } from './WarningIcon'
-export { SparklesIcon } from './SparklesIcon'
-export { LanguageIcon } from './LanguageIcon'