+++ packages/smarthr-ui/src/components/Switch/Switch.tsx
@@ -62,7 +62,7 @@ type Props = InputHTMLAttributes<HTMLInputElement> & {
 }
 
 export const Switch = forwardRef<HTMLInputElement, Props>(
-  ({ children, dangerouslyLabelHidden, className, id, ...props }, ref) => {
+  ({ children, dangerouslyLabelHidden, className, id, ...rest }, ref) => {
     const defaultId = useId()
     const inputId = id || defaultId
 
@@ -78,13 +78,13 @@ export const Switch = forwardRef<HTMLInputElement, Props>(
     }, [className])
 
     return (
-      <Cluster align="center">
+      <Cluster align="center" className="shr-relative">
         <MemoizedLabel htmlFor={inputId} dangerouslyLabelHidden={dangerouslyLabelHidden}>
           {children}
         </MemoizedLabel>
         <span className={classNames.wrapper}>
           <input
-            {...props}
+            {...rest}
             type="checkbox"
             role="switch"
             id={inputId}