+++ packages/smarthr-ui/src/components/Tooltip/Tooltip.tsx
@@ -38,7 +38,7 @@ const subscribeFullscreenChange = (callback: () => void) => {
 const getFullscreenElement = () => document.fullscreenElement
 const getFullscreenElementOnSSR = () => null
 
-type Props = PropsWithChildren<{
+type AbstractProps = PropsWithChildren<{
   /** ツールチップ内に表示するメッセージ */
   message: ReactNode
   /** ツールチップを表示する対象のタイプ。アイコンの場合は `icon` を指定する */
@@ -50,11 +50,12 @@ type Props = PropsWithChildren<{
   /** ツールチップを内包要素に紐付けるかどうか */
   ariaDescribedbyTarget?: 'wrapper' | 'inner'
 }>
-type ElementProps = Omit<ComponentProps<'span'>, keyof Props | 'aria-describedby'>
+type Props = AbstractProps & Omit<ComponentProps<'span'>, keyof AbstractProps | 'aria-describedby'>
 
 const classNameGenerator = tv({
   base: [
     'smarthr-ui-Tooltip',
+    'shr-relative',
     'shr-inline-block shr-max-w-full shr-align-bottom',
     'focus-visible:shr-focus-indicator',
   ],
@@ -65,7 +66,7 @@ const classNameGenerator = tv({
   },
 })
 
-export const Tooltip: FC<Props & ElementProps> = ({
+export const Tooltip: FC<Props> = ({
   message,
   children,
   triggerType,
@@ -79,7 +80,7 @@ export const Tooltip: FC<Props & ElementProps> = ({
   onTouchEnd,
   onFocus,
   onBlur,
-  ...props
+  ...rest
 }) => {
   const [portalRoot, setPortalRoot] = useState<Element | null>(null)
   const [isVisible, setIsVisible] = useState(false)
@@ -201,7 +202,7 @@ export const Tooltip: FC<Props & ElementProps> = ({
   return (
     // eslint-disable-next-line jsx-a11y/no-static-element-interactions,smarthr/a11y-delegate-element-has-role-presentation
     <span
-      {...props}
+      {...rest}
       ref={ref}
       tabIndex={tabIndex}
       aria-describedby={isInnerTarget ? undefined : messageId}